fix: prepare the engine primitive to handle distinct aggregation
